feat: cleanup and update libraries
