+++ .formatter.exs
@@ -1,4 +1,5 @@
 # Used by "mix format"
 [
+  import_deps: [:tesla],
   inputs: ["mix.exs", "{config,lib,test}/**/*.{ex,exs}"]
 ]
