+++ lib/microsoft/azure/storage/shared_access_signature.ex
@@ -1,6 +1,5 @@
 defmodule Microsoft.Azure.Storage.SharedAccessSignature do
   alias Microsoft.Azure.Storage
-  alias Microsoft.Azure.Storage.ApiVersion
   import Microsoft.Azure.Storage.Utilities, only: [add_to: 3, set_to_string: 2]
 
   # https://docs.microsoft.com/en-us/rest/api/storageservices/delegating-access-with-a-shared-access-signature
@@ -16,7 +15,6 @@ defmodule Microsoft.Azure.Storage.SharedAccessSignature do
     :start_time,
     :expiry_time,
     :resource,
-    :permissions,
     :ip_range,
     :protocol
   ]
@@ -56,7 +54,6 @@ defmodule Microsoft.Azure.Storage.SharedAccessSignature do
   def add_resource_type_object(v = %__MODULE__{target_scope: :account}),
     do: v |> add_to(:resource_type, :object)
 
-
   @resource_map %{
     # https://docs.microsoft.com/en-us/rest/api/storageservices/constructing-a-service-sas#specifying-the-signed-resource-blob-service-only
     container: "c",
@@ -69,7 +66,6 @@ defmodule Microsoft.Azure.Storage.SharedAccessSignature do
   def add_resource_blob_container(v = %__MODULE__{}), do: v |> add_to(:resource, :container)
 
   def add_resource_blob_blob(v = %__MODULE__{}), do: v |> add_to(:resource, :blob)
-
 
   @permissions_map %{
     read: "r",
@@ -159,7 +155,6 @@ defmodule Microsoft.Azure.Storage.SharedAccessSignature do
         ""
       ]
       |> Enum.join("\n")
-      |> IO.inspect(label: "stringToSign")
 
     signature =
       :crypto.hmac(:sha256, account_key |> Base.decode64!(), stringToSign)
@@ -168,53 +163,5 @@ defmodule Microsoft.Azure.Storage.SharedAccessSignature do
     values
     |> Map.put("sig", signature)
     |> URI.encode_query()
-  end
-
-  def sas1() do
-    new()
-    |> for_storage_account()
-    |> add_service_table()
-    |> add_service_queue()
-    |> add_service_queue()
-    |> add_service_queue()
-    |> add_resource_type_service()
-    |> add_resource_type_object()
-    |> add_permission_read()
-    |> ip_range("168.1.5.60-168.1.5.70")
-    # |> for_blob_service()
-    |> start_time(Timex.now())
-    |> expiry_time(Timex.now() |> Timex.add(Timex.Duration.from_hours(1)))
-    |> protocol("https")
-  end
-
-  def demo() do
-    sas1()
-    |> sign(%Storage{
-      cloud_environment_suffix: "core.windows.net",
-      account_name: "SAMPLE_STORAGE_ACCOUNT_NAME" |> System.get_env(),
-      account_key: "SAMPLE_STORAGE_ACCOUNT_KEY" |> System.get_env()
-    })
-    |> URI.decode_query()
-  end
-
-  def d2() do
-    new()
-    |> service_version(ApiVersion.get_api_version(:storage))
-    |> for_storage_account()
-    |> add_service_blob()
-    |> add_resource_type_container()
-    |> add_resource_blob_container()
-    |> add_permission_read()
-    |> add_permission_process()
-    |> add_permission_list()
-    |> start_time(Timex.now())
-    |> expiry_time(Timex.now()
-    |> Timex.add(Timex.Duration.from_days(100)))
-    |> sign(%Storage{
-          cloud_environment_suffix: "core.windows.net",
-          account_name: "SAMPLE_STORAGE_ACCOUNT_NAME" |> System.get_env(),
-          account_key: "SAMPLE_STORAGE_ACCOUNT_KEY" |> System.get_env()
-        }
-      )
   end
 end
