--- tesla_bugfix_when_no_connectivity.md
@@ -1,10 +1,0 @@
-
-- `tesla\lib\tesla\middleware\core.ex`
-
-
-```elixir
-defmodule Tesla.Middleware.Normalize do
-  def normalize(:error) do
-    raise %Tesla.Error{message: "unknown adapter error", reason: "unknown error"}
-  end
-```
